+++ projects/_01_ingest/cqc_api/fargate/cqc_locations_1_delta_api_download.py
@@ -5,6 +5,7 @@ import os
 import sys
 from datetime import date
 from datetime import datetime as dt
+from datetime import timedelta
 from typing import Generator
 
 import polars as pl
@@ -17,21 +18,20 @@ from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns as ColNames,
 )
 
-
-class InvalidTimestampArgumentError(Exception):
-    pass
-
-
 SECRET_ID = os.environ.get("CQC_SECRET_NAME", "")
 AWS_REGION = os.environ.get("AWS_REGION", "")
 CQC_OBJECT_TYPE = "locations"
 CQC_ORG_TYPE = "location"
 
 
-def main(destination: str, start_timestamp: str, end_timestamp: str) -> None:
+def main(
+    destination: str,
+    end_timestamp: str,
+    previous_days_to_capture: int = cqc.days_to_rollback_start_timestamp,
+) -> None:
     """
     This function performs the following steps:
-    1. Validates the provided start and end timestamps.
+    1. Subtracts a number of days from end_timestamp to create a start date.
     2. Retrieves the CQC API subscription key from AWS Secrets Manager.
     3. Calls the CQC API to fetch updated location objects within the specified
        time range.
@@ -44,16 +44,14 @@ def main(destination: str, start_timestamp: str, end_timestamp: str) -> None:
     Args:
         destination (str): The S3 path or local file path where the processed
             Parquet file will be written.
-        start_timestamp (str): The ISO 8601 formatted string representing the
-            start of the data retrieval period (e.g., '2023-01-01T00:00:00Z').
         end_timestamp (str): The ISO 8601 formatted string representing the
             end of the data retrieval period (e.g., '2023-01-31T23:59:59Z').
+        previous_days_to_capture (int): Integer of days before end_timestamp.
 
     Return:
         None
 
     Raises:
-        InvalidTimestampArgumentError: If `start_timestamp` is after `end_timestamp`.
         FileNotFoundError: If the function is unable to write the Parquet
             file to the specified `destination`.
         Exception: For any other unspecified errors that occur during API
@@ -62,19 +60,16 @@ def main(destination: str, start_timestamp: str, end_timestamp: str) -> None:
     try:
         destination = destination if destination[-1] == "/" else f"{destination}/"
 
-        start_dt = dt.fromisoformat(start_timestamp.replace("Z", ""))
         end_dt = dt.fromisoformat(end_timestamp.replace("Z", ""))
-
-        if start_dt > end_dt:
-            raise InvalidTimestampArgumentError(
-                "start_timestamp is after end_timestamp"
-            )
+        start_dt = end_dt - timedelta(days=previous_days_to_capture)
 
         print(f'Getting SecretID "{SECRET_ID}"')
         secret = get_secret(secret_name=SECRET_ID, region_name=AWS_REGION)
         cqc_api_primary_key_value: str = json.loads(secret)["Ocp-Apim-Subscription-Key"]
 
-        print("Collecting locations with changes from API")
+        print(
+            f"Collecting locations with changes from API between {start_dt} and {end_dt}"
+        )
 
         api_generator: Generator[dict, None, None] = cqc.get_updated_objects(
             object_type=CQC_OBJECT_TYPE,
@@ -106,10 +101,6 @@ def main(destination: str, start_timestamp: str, end_timestamp: str) -> None:
         utils.write_to_parquet(df_unique, destination)
         return None
 
-    except InvalidTimestampArgumentError as e:
-        print(f"ERROR: {e}")
-        print(f"ERROR: {sys.argv}")
-        raise
     except FileNotFoundError as e:
         print(f"ERROR: {e}")
         print(f"ERROR: {sys.argv}")
@@ -125,10 +116,16 @@ if __name__ == "__main__":
             "--destination_prefix",
             "Source s3 directory for parquet CQC locations dataset",
         ),
-        ("--start_timestamp", "Start timestamp for location changes"),
-        ("--end_timestamp", "End timestamp for location changes"),
+        (
+            "--end_timestamp",
+            "End timestamp for location changes",
+        ),
+        (
+            "--previous_days_to_capture",
+            "Number of days prior to end timestamp as a string",
+        ),
     )
-    print(f"Running job from {args.start_timestamp} to {args.end_timestamp}")
+    print(f"Running cqc locations delta api download job")
 
     date_today = date.today()
     destination = utils.generate_s3_dir(
@@ -138,4 +135,5 @@ if __name__ == "__main__":
         date=date_today,
         version="3.1.1",
     )
-    main(destination, args.start_timestamp, args.end_timestamp)
+    previous_days_to_capture = int(args.previous_days_to_capture)
+    main(destination, args.end_timestamp, previous_days_to_capture)